Refector: 레이아웃 파일 구분 및 디렉토리로 위계 구분
헤더와 푸터에서 style 분리하고, 디렉토리 위계 구분.
Layout의 MainSection의 flex 정렬하는 css 설정 삭제.
