+++ src/components/Layout/index.tsx
@@ -17,10 +17,6 @@ const Layout = () => {
 
 const MainSection = styled.main`
   min-height: calc(100vh - 350px);
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
 `;
 
 export default Layout;
